requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
SQLAlchemy==1.4.52
pandas==2.1.4
datetime
